Refactor home page into modular sections
